+++ pkg/registry/registry_test.go
@@ -29,24 +29,3 @@ func TestMain(m *testing.M) {
 func TestNew(t *testing.T) {
 	assert.NotNil(t, rc)
 }
-
-func TestTags(t *testing.T) {
-	assert.NotNil(t, rc)
-
-	image, err := registry.ParseImage(registry.ParseImageOptions{
-		Name: "crazymax/diun:3.0.0",
-	})
-	if err != nil {
-		t.Error(err)
-	}
-
-	tags, err := rc.Tags(registry.TagsOptions{
-		Image: image,
-	})
-	if err != nil {
-		t.Error(err)
-	}
-
-	assert.True(t, tags.Total > 0)
-	assert.True(t, len(tags.List) > 0)
-}
